split out json readlines to sep test class
